added check to see if user exists

+++ src/main/java/com/mattbroph/restfulapi/UserService.java
@@ -4,7 +4,10 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mattbroph.entity.User;
 import com.mattbroph.persistence.GenericDao;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.*;
 import javax.ws.rs.core.Response;
 import java.time.LocalDate;
@@ -16,21 +19,19 @@ import java.util.Map;
 
 public class UserService {
 
+    // Class instance variables
+    private final GenericDao userDao = new GenericDao(User.class);
+    private final Logger logger = LogManager.getLogger(this.getClass());
+
+
+    // Call http://localhost:8080/userDisplay/services/users
     @GET
     @Produces("application/json")
     public Response getUsers() throws JsonProcessingException {
 
-
-        // Call http://localhost:8080/userDisplay/services/users
-        GenericDao userDao = new GenericDao(User.class);
         List<User> users = userDao.getAll();
         ObjectMapper objectMapper = new ObjectMapper();
-        /*
-         * Referenced site below on how to convert List to JSON using Jackson
-         * https://www.baeldung.com/java-converting-list-to-json-array
-         * Reference site below to handle infinite loop by using
-         * https://stackoverflow.com/questions/3325387/infinite-recursion-with-jackson-json-and-hibernate-jpa-issue
-         */
+
         String usersJson = objectMapper.writeValueAsString(users);
         return Response.status(200).entity(usersJson).build();
     }
@@ -41,13 +42,16 @@ public class UserService {
     @Produces("application/json")
     public Response getUserFromId(@PathParam("id") int id) throws JsonProcessingException {
 
-        GenericDao userDao = new GenericDao(User.class);
-        User user = (User)userDao.getById(id);
         ObjectMapper objectMapper = new ObjectMapper();
+        User user = (User)userDao.getById(id);
+
+        // If user doesn't exist send them a 404
+        if (user == null) {
+            return Response.status(404).entity("User not found").build();
+        }
+
         String userJson = objectMapper.writeValueAsString(user);
-
         return Response.status(200).entity(userJson).build();
-
     }
 
     // Call http://localhost:8080/userDisplay/services/users/{idToDelete}
@@ -56,27 +60,27 @@ public class UserService {
     @Produces("application/json")
     public Response deleteUserWithId(@PathParam("id") int id) throws JsonProcessingException {
 
-        GenericDao userDao = new GenericDao(User.class);
         User userToDelete = (User)userDao.getById(id);
+
+        // If user doesn't exist send them a 404
+        if (userToDelete == null) {
+            return Response.status(404).entity("User not found").build();
+        }
+
+        // Delete the user
         userDao.delete(userToDelete);
 
         String successResponse = "{0}";
         return Response.status(200).entity(successResponse).build();
     }
+
     // Call http://localhost:8080/userDisplay/services/users
     @POST
     @Consumes("application/json")
     @Produces("application/json")
     public Response createUser(String userJson) throws JsonProcessingException {
 
-        GenericDao userDao = new GenericDao(User.class);
-
         ObjectMapper mapper = new ObjectMapper();
-
-        /*
-        * Reference section 3.5 Creating Java Map From JSON String
-        * https://www.baeldung.com/jackson-object-mapper-tutorial
-         */
         Map<String, String> map = mapper.readValue(userJson, Map.class);
 
         String firstName = map.get("firstName");
@@ -99,9 +103,7 @@ public class UserService {
     @Produces("application/json")
     public Response createUser(String userJson, @PathParam("id") int id) throws JsonProcessingException {
 
-        GenericDao userDao = new GenericDao(User.class);
         User userToUpdate = (User)userDao.getById(id);
-
         ObjectMapper mapper = new ObjectMapper();
 
         // Get the values from the json in the request body
@@ -120,15 +122,25 @@ public class UserService {
         LocalDate dateOfBirth = LocalDate.parse(map.get("dateOfBirth"));
         userToUpdate.setDateOfBirth(dateOfBirth);
 
-        // Update the user in the DB
+        // Update the user in the database
         userDao.update(userToUpdate);
 
         String successResponse = "{0}";
-
         return Response.status(200).entity(successResponse).build();
-
     }
 
+
+
+    /*
+     * Referenced site below on how to convert List to JSON using Jackson
+     * https://www.baeldung.com/java-converting-list-to-json-array
+     * Reference site below to handle infinite loop by using
+     * https://stackoverflow.com/questions/3325387/infinite-recursion-with-jackson-json-and-hibernate-jpa-issue
+     */
+    /*
+     * Reference section 3.5 Creating Java Map From JSON String
+     * https://www.baeldung.com/jackson-object-mapper-tutorial
+     */
 
 
 }
